+++ extended_mypy_django_plugin/plugin/_known_annotations.py
@@ -1,7 +1,10 @@
 import enum
 
 
+class KnownClasses(enum.Enum):
+    CONCRETE = "extended_mypy_django_plugin.annotations.Concrete"
+
+
 class KnownAnnotations(enum.Enum):
     CONCRETE = "extended_mypy_django_plugin.annotations.Concrete"
-    CONCRETE_QUERYSET = "extended_mypy_django_plugin.annotations.ConcreteQuerySet"
     DEFAULT_QUERYSET = "extended_mypy_django_plugin.annotations.DefaultQuerySet"