+++ extended_mypy_django_plugin/plugin/_plugin.py
@@ -1,4 +1,6 @@
+import enum
 import sys
+from collections.abc import Callable
 from typing import Generic
 
 from mypy.checker import TypeChecker
@@ -141,8 +143,10 @@ class ExtendedMypyStubs(main.NewSemanalDjangoPlugin):
     @_hook.hook
     class get_dynamic_class_hook(Hook[DynamicClassDefContext, None]):
         """
-        This is used to find ``Concrete.type_var`` and turn that into a ``TypeVar``
-        representing each Concrete class of the abstract model provided.
+        This is used to find special methods on the ``Concrete`` class and do appropriate actions.
+
+        For ``Concrete.type_var`` we turn the result into a ``TypeVar`` that can only be one of
+        the concrete descendants of the specified class.
 
         So say we find::
 
@@ -153,27 +157,35 @@ class ExtendedMypyStubs(main.NewSemanalDjangoPlugin):
             T_Child = TypeVar("T_Child", Child1, Child2, Child3)
         """
 
+        class KnownConcreteMethods(enum.Enum):
+            type_var = "type_var"
+
+        method_name: KnownConcreteMethods
+
         def choose(self) -> bool:
             class_name, _, method_name = self.fullname.rpartition(".")
-            if method_name == "type_var":
+            try:
+                self.method_name = self.KnownConcreteMethods(method_name)
+            except ValueError:
+                return False
+            else:
                 info = self.plugin._get_typeinfo_or_none(class_name)
-                if info and info.has_base(ExtendedMypyStubs.Annotations.CONCRETE.value):
-                    return True
-
-            return False
+                return bool(info and info.has_base(_known_annotations.KnownClasses.CONCRETE.value))
 
         def run(self, ctx: DynamicClassDefContext) -> None:
             assert isinstance(ctx.api, SemanticAnalyzer)
 
             sem_analyzing = actions.SemAnalyzing(self.store, api=ctx.api)
 
-            return sem_analyzing.transform_type_var_classmethod(ctx)
+            if self.method_name is self.KnownConcreteMethods.type_var:
+                return sem_analyzing.transform_type_var_classmethod(ctx)
+            else:
+                assert_never(self.method_name)
 
     @_hook.hook
     class get_type_analyze_hook(Hook[AnalyzeTypeContext, MypyType]):
         """
-        Resolve classes annotated with ``Concrete``, ``ConcreteQuerySet`` and
-        ``DefaultQuerySet``.
+        Resolve classes annotated with ``Concrete`` or ``DefaultQuerySet``.
         """
 
         def choose(self) -> bool:
@@ -193,9 +205,6 @@ class ExtendedMypyStubs(main.NewSemanalDjangoPlugin):
 
             if name is Known.CONCRETE:
                 method = type_analyzer.find_concrete_models
-
-            elif name is Known.CONCRETE_QUERYSET:
-                method = type_analyzer.find_concrete_querysets
 
             elif name is Known.DEFAULT_QUERYSET:
                 method = type_analyzer.find_default_queryset
@@ -226,7 +235,7 @@ class ExtendedMypyStubs(main.NewSemanalDjangoPlugin):
                 ctx, resolve_manager_method_from_instance=resolve_manager_method_from_instance
             )
 
-    class SharedCallableHookLogic:
+    class SharedAnnotationHookLogic:
         """
         Shared logic for modifying the return type of methods and functions that use a concrete
         annotation with a type variable.
@@ -272,16 +281,22 @@ class ExtendedMypyStubs(main.NewSemanalDjangoPlugin):
             return type_checking.modify_return_type(ctx)
 
     class _get_method_or_function_hook(Hook[MethodContext | FunctionContext, MypyType]):
+        runner: Callable[[MethodContext | FunctionContext], MypyType | None]
+
         def extra_init(self) -> None:
-            self.shared_logic = self.plugin.SharedCallableHookLogic(
+            self.shared_logic = self.plugin.SharedAnnotationHookLogic(
                 fullname=self.fullname, plugin=self.plugin
             )
 
         def choose(self) -> bool:
-            return self.shared_logic.choose()
+            if self.shared_logic.choose():
+                self.runner = self.shared_logic.run
+                return True
+            else:
+                return False
 
         def run(self, ctx: FunctionContext | MethodContext) -> MypyType:
-            result = self.shared_logic.run(ctx)
+            result = self.runner(ctx)
             if result is not None:
                 return result