+++ extended_mypy_django_plugin/plugin/actions/_type_analyze.py
@@ -1,11 +1,14 @@
+from collections.abc import Iterator, Sequence
+
 from mypy.nodes import TypeInfo
 from mypy.semanal import SemanticAnalyzer
 from mypy.typeanal import TypeAnalyser
 from mypy.types import (
     AnyType,
     Instance,
+    ProperType,
     TypeOfAny,
-    TypeVarType,
+    TypeType,
     UnboundType,
     UnionType,
     get_proper_type,
@@ -25,45 +28,96 @@ class TypeAnalyzing:
         self.sem_api = sem_api
         self.store = store
 
-    def find_concrete_models(self, unbound_type: UnboundType) -> MypyType:
+    def _flatten_union(self, typ: ProperType) -> Iterator[ProperType]:
+        if isinstance(typ, UnionType):
+            for item in typ.items:
+                yield from self._flatten_union(get_proper_type(item))
+        else:
+            yield typ
+
+    def _analyze_first_type_arg(
+        self, unbound_type: UnboundType
+    ) -> tuple[bool, Sequence[Instance] | None]:
         args = unbound_type.args
         type_arg = get_proper_type(self.api.analyze_type(args[0]))
 
-        if not isinstance(type_arg, Instance):
-            return unbound_type
+        is_type: bool = False
+        if isinstance(type_arg, TypeType):
+            is_type = True
+            type_arg = type_arg.item
 
-        concrete = tuple(
-            self.store.retrieve_concrete_children_types(
-                type_arg.type, self.lookup_info, self.sem_api.named_type_or_none
+        if isinstance(type_arg, AnyType):
+            self.api.fail("Tried to use concrete annotations on a typing.Any", unbound_type)
+            return False, None
+
+        if not isinstance(type_arg, Instance | UnionType):
+            return False, None
+
+        if isinstance(type_arg, Instance):
+            type_arg = UnionType((type_arg,))
+
+        all_types = list(self._flatten_union(type_arg))
+        all_instances: list[Instance] = []
+        not_all_instances: bool = False
+        for item in all_types:
+            if not isinstance(item, Instance):
+                self.sem_api.fail(
+                    f"Expected to operate on specific classes, got a {item.__class__.__name__}: {item}",
+                    unbound_type,
+                )
+                not_all_instances = True
+            else:
+                all_instances.append(item)
+
+        if not_all_instances:
+            return False, None
+
+        concrete: list[Instance] = []
+        names = ", ".join([item.type.fullname for item in all_instances])
+
+        for item in all_instances:
+            concrete.extend(
+                self.store.retrieve_concrete_children_types(
+                    item.type, self.lookup_info, self.sem_api.named_type_or_none
+                )
             )
-        )
+
         if not concrete:
             if self.sem_api.final_iteration:
-                self.api.fail(
-                    f"No concrete models found for {type_arg.type.fullname}", unbound_type
-                )
-                return AnyType(TypeOfAny.from_error)
+                self.api.fail(f"No concrete models found for {names}", unbound_type)
+                return False, None
             else:
                 self.sem_api.defer()
-                return unbound_type
+                return False, None
 
-        return UnionType(concrete)
+        return is_type, tuple(concrete)
 
-    def find_concrete_querysets(self, unbound_type: UnboundType) -> MypyType:
-        args = unbound_type.args
-        type_arg = get_proper_type(self.api.analyze_type(args[0]))
+    def _make_union(
+        self, is_type: bool, instances: Sequence[Instance]
+    ) -> UnionType | Instance | TypeType:
+        items: Sequence[UnionType | TypeType | Instance]
 
-        if not isinstance(type_arg, Instance):
-            return UnionType(())
+        if is_type:
+            items = [item if isinstance(item, TypeType) else TypeType(item) for item in instances]
+        else:
+            items = instances
 
-        concrete = tuple(
-            self.store.retrieve_concrete_children_types(
-                type_arg.type, self.lookup_info, self.sem_api.named_type_or_none
-            )
-        )
-        if not concrete:
-            self.api.fail(f"No concrete models found for {type_arg.type.fullname}", unbound_type)
-            return AnyType(TypeOfAny.from_error)
+        if len(items) == 1:
+            return items[0]
+        else:
+            return UnionType(tuple(items))
+
+    def find_concrete_models(self, unbound_type: UnboundType) -> MypyType:
+        is_type, concrete = self._analyze_first_type_arg(unbound_type)
+        if concrete is None:
+            return unbound_type
+
+        return self._make_union(is_type, concrete)
+
+    def find_default_queryset(self, unbound_type: UnboundType) -> MypyType:
+        is_type, concrete = self._analyze_first_type_arg(unbound_type)
+        if concrete is None:
+            return unbound_type
 
         try:
             querysets = tuple(self.store.realise_querysets(UnionType(concrete), self.lookup_info))
@@ -72,35 +126,9 @@ class TypeAnalyzing:
             return AnyType(TypeOfAny.from_error)
         except _store.UnionMustBeOfTypes:
             self.api.fail("Union must be of instances of models", unbound_type)
-            return AnyType(TypeOfAny.from_error)
-        else:
-            return UnionType(querysets)
-
-    def find_default_queryset(self, unbound_type: UnboundType) -> MypyType:
-        args = unbound_type.args
-        type_arg = get_proper_type(self.api.analyze_type(args[0]))
-
-        if isinstance(type_arg, AnyType):
-            self.api.fail("Can't get default query set for Any", unbound_type)
-            return unbound_type
-
-        if isinstance(type_arg, TypeVarType):
-            return unbound_type
-
-        if not isinstance(type_arg, Instance | UnionType):
-            self.api.fail("Default queryset needs a class to find for", unbound_type)
-            return unbound_type
-
-        try:
-            querysets = tuple(self.store.realise_querysets(type_arg, self.lookup_info))
-        except _store.RestartDmypy as err:
-            self.api.fail(f"You probably need to restart dmypy: {err}", unbound_type)
-            return AnyType(TypeOfAny.from_error)
-        except _store.UnionMustBeOfTypes:
-            self.api.fail("Union must be of instances of models", unbound_type)
             return unbound_type
         else:
-            return UnionType(querysets)
+            return self._make_union(is_type, querysets)
 
     def lookup_info(self, fullname: str) -> TypeInfo | None:
         instance = self.sem_api.named_type_or_none(fullname)